April 18 beginning of class
I added interactive points to each chart to each of the charts and made charts (both total and 90) for for goalkeeping stats. Also made tables for each stat and began to load in descriptions of the variables and added them to the charts so when you hover over a variable it displays a description.